+++ drivers/echo-speaks-device.groovy
@@ -25,8 +25,8 @@ import java.text.SimpleDateFormat
 //************************************************
 //*               STATIC VARIABLES               *
 //************************************************
-@Field static final String devVersionFLD  = '4.2.0.2'
-@Field static final String devModifiedFLD = '2022-04-18'
+@Field static final String devVersionFLD  = '4.2.0.6'
+@Field static final String devModifiedFLD = '2022-05-05'
 @Field static final String sNULL          = (String)null
 @Field static final String sBLANK         = ''
 @Field static final String sSPACE         = ' '
@@ -131,7 +131,7 @@ metadata {
         command "playCalendarTomorrow", [[name: "Set Volume", type: "NUMBER", description: "Sets the volume before playing the message"],[name: "Restore Volume", type: "NUMBER", description: "Restores the volume after playing the message"]]
         command "playCalendarNext", [[name: "Set Volume", type: "NUMBER", description: "Sets the volume before playing the message"],[name: "Restore Volume", type: "NUMBER", description: "Restores the volume after playing the message"]]
         command "stopAllDevices"
-        command "noOp", [[name: "Internal Command Only... Does nothing here...", description: "Internal Command Only... Does nothing Here..."]]
+        command "noOp"
 
         if(!isZone()) {
             command "searchMusic", [[name: "Music Search Phrase*", type: "STRING", description: "Enter the artist, song, playlist, etc."], [name: "Music Provider*", type: "ENUM", constraints: ["AMAZON_MUSIC", "APPLE_MUSIC", "TUNEIN", "PANDORA", "SIRIUSXM", "SPOTIFY", "I_HEART_RADIO", "CLOUDPLAYER"], description: "Select One of these Music Providers to use."], [name: "Set Volume", type: "NUMBER", description: "Sets the volume before playing"],[name: "Sleep Time", type: "NUMBER", description: "Sleep time in seconds"]]
@@ -324,7 +324,7 @@ Boolean isCommandTypeAllowed(String type, Boolean noLogs=false, Boolean ignoreDo
     if(!(String)state.deviceType) { if(!noLogs) { logWarn("DeviceType State Value Missing: ${(String)state.deviceType}", true) }; return false }
     if(!(String)state.deviceOwnerCustomerId) { if(!noLogs) { logWarn("OwnerCustomerId State Value Missing: ${(String)state.deviceOwnerCustomerId}", true) }; return false }
 
-    Boolean isOnline = (device?.currentValue("onlineStatus") == "online")
+    Boolean isOnline = settings.ignoreHealth!=false || device?.currentValue("onlineStatus") == "online"
     if(!isOnline) {
         if(!noLogs) { logWarn("Commands NOT Allowed! Device is currently (OFFLINE) | Type: (${type})", true) }
         triggerDataRrshF("found offline" )
@@ -2791,12 +2791,13 @@ void updateLevel(oldvolume, newvolume) {
     }// else logWarn("Uh-Oh... UpdateLevel without any values")
 }
 
-private void speechCmd(Map cmdMap=[:], Boolean parallel=false) {
+private void speechCmd(Map icmdMap=[:], Boolean parallel=false) {
+    Map cmdMap=[:]+icmdMap
     if(!cmdMap) { logError("speechCmd | Error | cmdMap is missing"); return }
     String healthStatus = getHealthStatus()
-    if(settings.ignoreHealth == false && !(healthStatus in ["ACTIVE", "ONLINE"])) { logWarn("speechCmd Ignored... Device is OFFLINE", true); return }
+    if((Boolean)settings.ignoreHealth == false && !(healthStatus in ["ACTIVE", "ONLINE"])) { logWarn("speechCmd Ignored... Device is OFFLINE", true); return }
 
-    if(settings.logTrace){
+    if((Boolean)settings.logTrace){
         String tr = "speechCmd (${cmdMap.cmdDesc}) | Msg: ${fixLg(cmdMap.message.toString())}"
         tr += cmdMap.newVolume ? " | SetVolume: (${cmdMap.newVolume})" :sBLANK
         tr += cmdMap.oldVolume ? " | Restore Volume: (${cmdMap.oldVolume})" :sBLANK
@@ -3163,6 +3164,7 @@ static String getObjType(obj) {
     else if(obj instanceof BigDecimal) {return "BigDecimal"}
     else if(obj instanceof Float) {return "Float"}
     else if(obj instanceof Byte) {return "Byte"}
+    else if(obj instanceof com.hubitat.app.DeviceWrapper)return 'Device'
     else { return "unknown"}
 }